third_party/WebKit/LayoutTests/fast/canvas/OffscreenCanvasTemp-invalid-args-expected.txt

Tests that the OffscreenCanvasTemp can handle invalid arguments

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS canvas1.width is setWidth-1
PASS canvas1.height is setHeight-1
PASS canvas1.width is 0
PASS canvas1.height is 0
PASS new OffscreenCanvasTemp(-1, -1) threw exception TypeError: Failed to construct 'OffscreenCanvasTemp': Value is outside the 'unsigned long' value range..
PASS canvas2.width is 0
PASS canvas2.height is 0
PASS canvas2.width is setWidth-1
PASS canvas2.height is setHeight-1
PASS canvas2.width = -1 threw exception TypeError: Failed to set the 'width' property on 'OffscreenCanvasTemp': Value is outside the 'unsigned long' value range..
PASS canvas2.height = -1 threw exception TypeError: Failed to set the 'height' property on 'OffscreenCanvasTemp': Value is outside the 'unsigned long' value range..
PASS canvas2.width = obj threw exception TypeError: Failed to set the 'width' property on 'OffscreenCanvasTemp': Value is not of type 'unsigned long'..
PASS canvas2.height = obj threw exception TypeError: Failed to set the 'height' property on 'OffscreenCanvasTemp': Value is not of type 'unsigned long'..
PASS new OffscreenCanvasTemp(obj, obj) threw exception TypeError: Failed to construct 'OffscreenCanvasTemp': Value is not of type 'unsigned long'..
PASS successfullyParsed is true

TEST COMPLETE
